move models to public directory
